radar & build paths

+++ .build/build-model.py
@@ -27,7 +27,7 @@ if not os.path.exists(mwmBuilder):
 			exec(open(buildIni).read())
 			
 		if not os.path.exists(mwmBuilder):
-			inScriptDir = scrip + "\\MwmBuilder.exe" # from script dir
+			inScriptDir = scriptDir + "\\MwmBuilder.exe" # from script dir
 			if (os.path.exists(inScriptDir)):
 				mwmBuilder = inScriptDir
 				
@@ -66,28 +66,27 @@ def emptyDir(l_dir):
 			if os.path.isfile(l_dir + "\\" + file):
 				os.remove(l_dir + "\\" + file)
 
-def copyWithExtension(l_from, l_to, l_ext):
+def copyWithExtension(l_from, l_to, l_ext, overwrite = False):
 	createDir(l_to)
 	os.chdir(l_from)
 	for file in os.listdir('.'):
 		if file.lower().endswith(l_ext.lower()):
-			shutil.copy2(file, l_to)
-
+			outFile = l_to + '\\' + os.path.basename(file)
+			if not os.path.exists(outFile) or overwrite:
+				shutil.copy2(file, l_to)
 
 # set up directories for MwmBuilder
-emptyDir(startDir + "\\" + input)
-emptyDir(startDir + "\\" + output)
 createDir(input)
 createDir(output)
 copyWithExtension(startDir, input, ".fbx")
-copyWithExtension(startDir, input, ".xml")
 copyWithExtension(startDir, input, ".hkt")
 copyWithExtension(startDir, output, ".mwm") # copying.mwm to output allows MwmBuilder to skip unchanged models
+copyWithExtension(startDir, input, ".xml")
 
 # run MwmBuilder
 mwmBuilderProcess = subprocess.Popen([mwmBuilder, "/s:" + input, "/o:" + output, "/l:" + startDir + "\\MwmBuilder.log"])
 mwmBuilderProcess.wait()
-copyWithExtension(output, startDir, ".mwm")
+copyWithExtension(output, startDir, ".mwm", True)
 
 # cannot delete directories, but we can empty them
 emptyDir(startDir + "\\" + input)
